rename files, reorg folder structure
