+++ src/controllers/api/conversation.ts
@@ -2,7 +2,7 @@ import { Router, Request, Response } from 'express';
 
 import JSONResponse from './utils/JSONResponse';
 import { statusCodes, errorMessages } from './utils/constants';
-import Conversation from '../models/Conversation';
+import Conversation from '../../models/Conversation';
 
 const createConversation = async (req: Request, res: Response): Promise<Response> => {
   const { conversation } = req.body;
