+++ src/server.ts
@@ -5,7 +5,7 @@ import socketIo from 'socket.io';
 dotenv.config();
 
 import app from './app';
-import addSocketHandlers from './controllers/socket';
+import addSocketHandlers from './controllers/socket/socket';
 
 const PORT = process.env.PORT || 8080;
 const server = http.createServer(app);
